Solution to Check Alphabet Case

+++ Check Alphabet Case@@6788dd63490637cfc44e1b92/code.c
@@ -7,6 +7,6 @@ char* welcome() {
 int main() {
     int a;
     scanf("%d",&a);
-    printf("%d",(int)a);
+    printf("%d",a);
     return 0;
 }
